code review
Seperated out the different terms by class (environment, model terms, final calculation).
Provided equation numbers to reference the paper directly.
Changed the calculation of Jmax to more closely align with the presentation in the manuscript (first calculate Jmax before calculating the JV ratio).
Changed the name of calc_vcmax_tresp_mult.R to reflect that the response is being applied to Jmax and Vcmax.
I suggest using arrows instead of equals for assignment in R. Although they often function equivalently, that is not always the case and the R style guide prohibits the use of '=' for assignment.